Use `Homebrew::CLI::Parser`
This helps us with consistent help output, parsing invalid options and
accessing options without using ARGV directly.

While we're here, use this chance to simplify the code around argument
parsing and different commands.
