--- cmd/brew-unalias.rb
@@ -1,1 +1,0 @@
-brew-alias.rb
